// src/app/core/interfaces/pokemonTypeName.interface.ts
export type PokemonTypeName = 
                                | 'normal'
                                | 'fighting'
                                | 'flying'
                                | 'poison'
                                | 'ground'
                                | 'rock'
                                | 'bug'
                                | 'ghost'
                                | 'steel'
                                | 'fire'
                                | 'water'
                                | 'grass'
                                | 'electric'
                                | 'psychic'
                                | 'ice'
                                | 'dragon'
                                | 'dark'
                                | 'fairy'
                                | 'stellar'
                                | 'unknown';